Simplified the IMSI catcher code
