+++ .gitignore
@@ -182,3 +182,4 @@ pyrightconfig.json
 venv
 .env
 test*
+.idea
