+++ common/gsmtaphdr.py
@@ -1,0 +1,28 @@
+import ctypes
+
+
+class GsmtapHdr(ctypes.BigEndianStructure):
+    _pack_ = 1
+    # Based on gsmtap_hdr structure in <grgsm/gsmtap.h> from gr-gsm
+    _fields_ = [
+        ("version", ctypes.c_ubyte),
+        ("hdr_len", ctypes.c_ubyte),
+        ("type", ctypes.c_ubyte),
+        ("timeslot", ctypes.c_ubyte),
+        ("arfcn", ctypes.c_uint16),
+        ("signal_dbm", ctypes.c_ubyte),
+        ("snr_db", ctypes.c_ubyte),
+        ("frame_number", ctypes.c_uint32),
+        ("sub_type", ctypes.c_ubyte),
+        ("antenna_nr", ctypes.c_ubyte),
+        ("sub_slot", ctypes.c_ubyte),
+        ("res", ctypes.c_ubyte),
+    ]
+
+    def __repr__(self):
+        return "%s(version=%d, hdr_len=%d, type=%d, timeslot=%d, arfcn=%d, signal_dbm=%d, snr_db=%d, frame_number=%d, sub_type=%d, antenna_nr=%d, sub_slot=%d, res=%d)" % (
+            self.__class__, self.version, self.hdr_len, self.type,
+            self.timeslot, self.arfcn, self.signal_dbm, self.snr_db,
+            self.frame_number, self.sub_type, self.antenna_nr, self.sub_slot,
+            self.res,
+        )
